Split BlockDodge and BlockDodgePanel back up.
